D3D12: Implement D3D12Context.blit()
This change introduces implementation of blit() functionality in D3D12
backend. D3D12 does not support blitting as an operation, so it had to
be implemented "the long way".

Depending on the scenario, the backend will opt for one of four paths:
- Fast path with non-MSAA source RT (== CopyTextureRegion)
- Fast path with MSAA source RT (== ResolveSubresourceRegion)
- Slow path with non-MSAA source RT (== quad render via BlitPS)
- Slow path with MSAA source RT (== ResolveSubresource + BlitPS)

It is assumed that we are always blitting to 1-sample destination
texture. This I believe was also the assumption of the blit()
implementation in D3D9 - StretchRect() only supported source MSAA texture.
But, if needed, with some testing this restriction could probably be
lifted (we'd always have to use the slow path then).

The choice between slow and fast path is done depending on source and
destination texture sizes - fast path is taken when destination box has
same width and height as source box, otherwise slow path is taken.

Fast path performs a simple Copy and/or Resolve call, while slow path
employs the newly added Blit Pixel Shader to resample the source RT to
destination dimensions. This also means, slow path has to preserve current
Rendering Context and potentially create an intermediate Texture (when source
RT has MSAA enabled) to resolve source RT to a 1-sample texture.

Some other changes and adjustments were made in the process to
accommodate this change + fix some very minor earlier mistakes.