RingContainer: Fix potential data overwriting
In some rare cases it was possible for RingContainer-s to incorrectly
detect the situation where the RingContainer is full. First, the
condition for "tail is behind head" was wrong, and then the check for
overflow would be incorrect. This could cause the area near head to be
treated as available and thus overwritten when preparing a Command List.
Overwriting could ex. happen when GPU is still executing a Command List
which uses some temporary Descriptors, which would cause device removal.

This adjustment to conditions ensures that:
- We only enter the "tail behind head" section when it actually is the
  case
- We will wait for next mid-frame checkpoint when Ring Container is
  "almost" full

Fixes occassional "Failed to Present on SwapChain" bug, leading to
infinite-loop of "Failed to Create SwapChain" messages.