Unify string-wstring conversion routines
Conversion from string to wstring and back was moved to D3D12Utils and
rest of the backend uses said conversion routines.

Current conversion code (using STL codecvt) will have to be replaced
once we upgrade the C++ compiler (newer C++ versions deprecated codecvt
in place of new routines) so this way it will be much easier to do so.