Mtl1PS: Fix addition of specular component
In situations where we turned out to not have any specular component to
add, specPower could be calculated as 0.0, which when calculating
specular contrubiton with pow() we would get 0^0, which results in -NaN.

Between HLSL3 and HLSL6 (or between FXC and DXC) there had to be a
change in how some math is handled by the shaders, because originally
D3D9 would just calculate the result to 0.0 contribution for specular.
Here we have to be more mindful, as -NaN propagates through light
equations and ends up -NaN-ing the resulting color. Problem was most
noticeable on Ensemble8 3DSphere demo, where "land" rendered incorrectly.

To fix the problem, specular power is checked in computeLight() and
specular contribution is only added when power is greater than 0.0.
Moreover, the computeLight for loop now only loops based on actual light
count (numShaderLights instead of numLights). That prevents unnecessary
calculation of light components which are potentially not initialized and
can also result in -NaN being propagated to result color, plus it
optimizes the shader a bit for lower light scenes.