Fix Depth Buffer clearing
In situations where we render a SubScene with depth testing enabled
Prism would often follow below pattern for clearing the RTT:
- set the RTT with depth test disabled
- clear the RTT
- re-set the RTT with depth test enabled

D3D12 assumed that when the depth testing is disabled then the Depth
Buffer should not be cleared even when it is present. This turned out to
not be the case, as D3D9 backend cleared the Depth Buffer even if the
testing was disabled. As such, in some scenarios (HelloViewOrder) we
missed a Depth Buffer clear which created visual bugs.

This change fixes above scenario and makes it so that Clear() can be
issued with clearDepth flag. If present and Depth Buffer has already
been created the Clear() will also issue a Depth Buffer clear.

Fixes HelloViewOrder demo not drawing the subscene properly and in
correct order.