Workaround issues with 3D rendering glitches
Adds a wait in RingContainer to stabilize the backend.

There is a bug somewhere which causes CPU-side to overwrite mid-frame
data before GPU consumed it fully. This causes flickering and can result
in device being removed. This patch mitigates this, however it also
reduces CPU-GPU parallelism, which impacts performance.

The actual bug needs more investigation, so for stabilization purposes
this workaround is temporarily added and will be removed when proper
solution will be found.