Add prism.d3d12.midframeWait property
Temporarily added, this property enables/disables waiting for GPU
right after Ring Container decides to flush the current Command List.
This allows to turn on/off the workaround for JDK-8356029.

This property is enabled by default, disabling it might cause glitches
and/or cause D3D12 device to be removed. It will be removed once
JDK-8356029 is properly fixed.