8356859: Rewrite Descriptor management to create-once-and-copy scheme
This change adds DescriptorAllocator which manages descriptor allocation
and multiple Heaps. This allows to allocate as many descriptors as the
system requires.

With the Allocator it is now possible to pre-allocate the most commonly
used SRV Descriptor for all Textures and use CopyDescriptorsSimple()
instead of creating a new one mid-frame, which should lower the CPU load
during rendering.