8356861: Fix performance of RenderPerfTest Blend examples
Prism's PrRenderer class tries to recognize whether current rendering
backend supports ShaderModel.SM3 Shaders and initializes hardware
filters and shaders if that is the case.

Because D3D12 introduced a new ShaderModel - SM6 - this was not taken
into account and PrRenderer ran all filter effects via the software/SSE
pipeline. This in turn caused frequent readPixels() calls to D3D12
renderer, which allocated a lot of Textures and RTTs, and flushed the
Command Queue every time a blended object was rendered.

Fixing this part of PrRenderer brings D3D12's performance to levels of
other tests which utilize textures.