RingContainer: Assume default size to be 3 times flush threshold
This change swaps the size and flushThreshold initialization parameters
of RingContainer and derived classes, assuming the size by default is 0.

With size 0 RingContainer calculates the size to be 3 times the flush
threshold which enables triple-buffering of mid-frame data. The only
exception is the Sampler Heap which has a hard limit of 2048 slots.