Make Ring Container resources configurable
This adds three properties managing three mid-frame containers. All of
them prefix with usual "prism.d3d12":
- .mainRingBufferThreshold - determines the size of Main Ring Buffer's
  flush threshold
- .constantRingBufferThreshold - determines the size of Constant Ring
  Buffer's flush threshold
- srvRingHeapThreshold - determines the size of CBV/SRV/UAV Ring
  Descriptor Heap's flush threshold

All three containers default to total size of 3 times the threshold.

This allows for changing the size of these containers without
recompiling the whole backend, which can be useful in optimizing the
backend on different hardware.