8368300: JavaFX D3D12: Optimize Blending operations
In some cases it is possible to avoid an RTT Clear() call. The
conditions that were estimated are:
- Attempting an RTT clear to RGBA 0.0f, 0.0f, 0.0f, 0.0f
- Following the clear by a draw call that:
  - Overlaps the RTT surface that was used
  - Uses SRC_OVER composite mode

In above cases it is possible to discard a Clear() call and instead draw
using SRC composite mode.

This change implements this behavior by delaying the Clear() call when it
is attempted with RGBA values set to all zeros and reattempting it right
before the Draw() call. If above conditions are met - correct composite
mode is used and quad area being rendered matches or exceeds RTTs dirty
region - the Clear() call is discarded and the Draw() is performed using
SRC composite mode, overwriting old RTT contents.

There were some situations (notably Ensemble8's Menu Bar and forward/back
buttons) which call a Clear() and then swap the RTT without rendering
anything, so the delayed Clear() call also has to be checked for and
recorded when switching RTTs.

This change brings Blend examples to matching/surpassing level of D3D
backend. It is worth noting that similar behavior is most probably
implemented in D3D9 drivers, as removing the affecting Clear() from
Prism's ImagePool.java@161 affected D3D12's performance while NOT
changing the performance of Blend perf tests on D3D backend.