8369117: JavaFX D3D12: Crash when entering fullscreen via ALT+ENTER
Alt+ENTER is a shortcut not officially supported by us. D3D12 uses DXGI
to get the Device, which injects the shortcut by default.

Added a call to DXGIFactory with appropriate flags which disable the
Alt+ENTER shortcut and tell DXGI to not make any window changes, as
JavaFX already handles those.