Exclude SwapChain's offscreen RTT from dirty bbox opts
JavaFX already tracks SwapChain dirty regions (and does it way better
and in a way more complex manner) which clashes with the simpler dirty
region algorithm existing in D3D12 backend.

Current implementation is still useful for temporary RTTs where there is
no optimization. This change adds a flag to disable the optimization
per-RTT and disables it for SwapChain's offscreen RTT while enabling it
for all other RTTs.